apps/backend/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "DreamManor - FastApi and Quasar"
authors = ["kochenov <kochenof@yandex.ru>"]
readme = "README.md"
#package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.0"}
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
gunicorn = "^21.2.0"
asyncpg = "^0.29.0"
fastapi-pagination = "^0.12.21"
celery = {extras = ["redis"], version = "^5.3.6"}
beautifulsoup4 = "^4.12.3"
selenium = "^4.18.1"
webdriver-manager = "^4.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[tool.mypy]
strict = true
exclude = ["venv", "alembic"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
